requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
python-multipart==0.0.6
jinja2==3.1.2
xhtml2pdf==0.2.11
python-dotenv==1.0.1